+++ ReleaseNotes.md
@@ -9,6 +9,7 @@
 ### Application V2 migration
 
 - Astrolabium - done
+- Activity Schedule - done
 - ChatMessage - removed jquery usage
 - Active effects config - Ars theme
 - All dialogs are now DialogV2